% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{mobility_app}
\alias{mobility_app}
\title{shiny app for mobility visualization}
\usage{
mobility_app(mobdata)
}
\arguments{
\item{mobdata}{a data.frame-like object with country_region and mobility measure data as used by plot_one}
}
\description{
shiny app for mobility visualization
}
