% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{plot_one}
\alias{plot_one}
\title{plot mobility traces over time for a single nation}
\usage{
plot_one(
  cntry = "Italy",
  datasource = system.file("csv/mobsel_aug15.csv.gz", package = "mobpack")
)
}
\arguments{
\item{cntry}{character(1)}

\item{datasource}{either character(1) in which case content of file named should be ingestible to
data.frame via vroom, or a data.frame-like object}
}
\description{
plot mobility traces over time for a single nation
}
\examples{
limdat = get_mob_sel()
plot_one(cntry = "India", datasource=limdat)
}
